fix(Reactions): Batch requests
A whitelisted method `batch` takes an array of requests and executes
all of them in a  single request, returning a list of responses.

On the client side, there a resource that calls this `batch` method and
it is debounced. User actions are kept in `changes` array and the batch
resource flushes them if there are no further changes in 1 second.

- also fix removeReaction bug